Add files via upload
The files with which the tests are done.

+++ s2_imu.h
@@ -17,7 +17,7 @@ class IMU_Class {
   IMU_Class();
 
   void IMU_initialization();
-  float IMU_loop();
+  int IMU_loop();
   int IMU_interaction();
